requirements.txt
torch
torchvision
torchaudio
ultralytics
numpy
Pillow

